api_for_test/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.5
django-cors-headers==3.13.0
django-extensions==3.2.1
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
PyJWT==2.6.0
pytz==2022.7
requests==2.28.1
responses==0.22.0
sqlparse==0.4.3
toml==0.10.2
types-toml==0.10.8.1
tzdata==2022.7
urllib3==1.26.13
whitenoise==6.2.0
